[Decode] change perf time base
change perf time base

+++ media_driver/agnostic/common/cm/cm_hal.h
@@ -1557,7 +1557,7 @@ typedef struct _CM_HAL_STATE
 
     CMRT_UMD::CSync             *criticalSectionDSH;
 
-    uint64_t                    tsFrequency;
+    uint32_t                    tsFrequency;
 
     bool                        forceKernelReload;
